feat: remove getDepositsAllocation and getTopUpAllocation methods
